requirements.txt
requests
matplotlib
networkx
pyyaml
scikit-learn
joblib
pandas
numpy
websocket-client
